refactor(types): move interfaces to dedicated files
