+++ components/admin/instruments/InstrumentForm.tsx
@@ -1,5 +1,5 @@
 import { useState, useEffect } from "react";
-import { InstrumentInfo } from "@/lib/types";
+import { InstrumentInfo } from "@/types";
 import { Input } from "@/components/ui/input";
 import { Button } from "@/components/ui/button";
 import { Textarea } from "@/components/ui/textarea";
