+++ types/preferences.ts
@@ -1,0 +1,6 @@
+export interface PreferenceValues {
+  notifications: boolean;
+  volume: boolean;
+  favorite: boolean;
+  globalMute?: boolean;
+}
